Feat: 구장 마커도 필터링 기능 추가
